gcp/requirements.txt
Flask==3.1.0
google-cloud-storage==2.9.0
tensorflow==2.10.0
Pillow==9.2.0
numpy==1.23.1